% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.r
\name{dimension}
\alias{dimension}
\alias{dim.adeba}
\title{Retrieve the dimension or dimensions of the dataset used to make the estimate}
\usage{
dimension(x)

\method{dim}{adeba}(x)
}
\arguments{
\item{x}{Density estimate.}
}
\value{
Integer scalar.

Integer vector.
}
\description{
Retrieve the dimension or dimensions of the dataset used to make the estimate
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
