% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_constants}
\alias{find_constants}
\title{Check for columns with constant values}
\usage{
find_constants(x)
}
\arguments{
\item{x}{A numeric matrix.}
}
\value{
A logical vector containing \code{TRUE} if all elements of a column
  are identical.
}
\description{
Check for columns with constant values
}
