% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.r
\name{is.rendered}
\alias{is.rendered}
\title{Determines if an estimate is rendered}
\usage{
is.rendered(object)
}
\arguments{
\item{object}{Density estimate.}
}
\value{
Logical.
}
\description{
Determines if an estimate is rendered
}
\seealso{
render
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
