% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_constant}
\alias{is_constant}
\title{Check if a vector contains only one unique value}
\usage{
is_constant(x)
}
\arguments{
\item{x}{A numeric vector with no missing values.}
}
\value{
A logical that is \code{TRUE} if all elements were identical,
  otherwise \code{FALSE}.
}
\description{
Check if a vector contains only one unique value
}

