% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.r
\name{iterate}
\alias{iterate}
\title{Fit parameters}
\usage{
iterate(object, ...)
}
\arguments{
\item{object}{Density estimate.}

\item{...}{Sent to the internal function \code{get_log_likelihood}.}
}
\value{
An ADEBA estimate with increased number of iterations.
}
\description{
This function finds appropriate values for alpha, using a fixed beta=0.5, and
calculates the estimate.
}
\details{
To identify a suitable range for alpha containing the bulk of its posterior
probability, a rough logarithmically spaced grid is searched.
The returned interval is searched again to yield more accurate endpoints,
and finally a fine linearly spaced grid is used for the final density
estimate.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
