% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{mono2png}
\alias{mono2png}
\title{Convert monochrome matrix to png array}
\usage{
mono2png(x, pal = c("#ff9933", "#ffff66"), alpha = TRUE)
}
\arguments{
\item{x}{Monochrome matrix (numeric with all elements in [0,1]).}

\item{pal}{Color palette.}

\item{alpha}{Whether to also use the monochrome information as alpha channel.}
}
\value{
An RGB or RGBA-array.
}
\description{
Convert monochrome matrix to png array
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
blend, mono2hex
}

