% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.adeba}
\alias{plot.adeba}
\alias{lines.adeba}
\alias{points.adeba}
\alias{contour.adeba}
\title{Plot density estimate}
\usage{
\method{plot}{adeba}(x, y, type = c("estimate", "data", "both"), ...,
  add = FALSE)

\method{lines}{adeba}(x, ...)

\method{points}{adeba}(x, ...)

\method{contour}{adeba}(x, ...)
}
\arguments{
\item{x}{Density estimate.}

\item{y}{Ignored, kept for S3 consistency.}

\item{type}{What type of plot to draw.}

\item{...}{Sent to \code{\link{plot}}, \code{\link{image}} or some other
base plotting function, depending on \code{type} and dimension of the
estimate.}

\item{add}{Whether to start a new plot (\code{FALSE}) or add to an existing
(\code{TRUE}).}
}
\description{
Plot density estimate
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
