% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot3d}
\alias{plot3d}
\title{Plot adeba estimate in 3d}
\usage{
plot3d(x, ...)
}
\arguments{
\item{x}{Object to plot.}

\item{...}{Sent to \code{\link[rgl]{plot3d}} of the \code{rgl} package.}
}
\description{
Using OpenGL. Re-exported from the \code{rgl} package, so please see
\code{\link[rgl]{plot3d}} for details.
}

