% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_alpha}
\alias{plot_alpha}
\alias{plot_beta}
\title{Plot the parameters and posterior}
\usage{
plot_alpha(x, ...)

plot_beta(x, ...)
}
\arguments{
\item{x}{ADEBA estimate.}

\item{...}{Ignored, kept for S3 consistency.}
}
\description{
\code{plot_beta} plots the total posterior mass for each value of the beta
parameter (integrated over alpha). \code{plot_alpha} plots the normalized
posterior of alpha given different beta values. The normalization adjusts the
posterior based on how densely the alpha values were sampled for each beta
value. This is a requirement for the integral approximation to hold.
}

