% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.r
\name{predict.adeba}
\alias{predict.adeba}
\title{Predict PDF at selected points}
\usage{
\method{predict}{adeba}(object, newx, ...)
}
\arguments{
\item{object}{Density estimate.}

\item{newx}{New data points to predict. Can either be a dataset of the same
dimension as was used to make the estimate, or a point grid in list form
(see \code{\link{render}} for details).}

\item{...}{Ignored, kept for S3 consistency.}
}
\value{
A vector if \code{newx} is a matrix or array, or an array if
  \code{newx}.
}
\description{
Predict PDF at selected points
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
