% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.r
\name{radeba}
\alias{radeba}
\title{Sample from ADEBA estimate}
\usage{
radeba(n, object)
}
\arguments{
\item{n}{Number of examples to sample.}

\item{object}{Density estimate.}
}
\description{
Works just like \code{\link{rnorm}}, \code{\link{runif}}, etc.
}
\examples{
data(faithful)
f <- adeba(faithful$eruptions, adaptive=FALSE)
f.eruptions <- radeba(1e5, f)
hist(f.eruptions, breaks=100, col="skyblue", probability=TRUE)
lines(f, lwd=3)
rug(faithful$eruptions)
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

