% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.r
\name{render}
\alias{is.rendered}
\alias{render}
\title{Render and store the estimate}
\usage{
render(object, grid)

is.rendered(object)
}
\arguments{
\item{object}{Density estimate.}

\item{grid}{A list of numerical vectors corresponding to the dimensions of
the estimate.}
}
\value{
Nothing, the object is modified in place.
}
\description{
Render and store the estimate
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

